+++ src/service/logService/index.ts
@@ -1,0 +1,2 @@
+export * from './getGoalLogs';
+export * from './addLogForGoal';
